requirements.txt
Django==1.2.5
MySQL-python==1.2.3
South==0.7.3
django-contact-form==0.3.1
django-debug-toolbar==0.8.5
django-picklefield==0.1.9
django-profiles==0.2
django-registration==0.8-alpha-1
ipython==0.11
python-dateutil==1.5
readline==6.2.0
tweepy==1.7.1
wsgiref==0.1.2
